Added some very ugly deprecation warnings to some files.
These files represent (to the best of my knowledge) some of our most common parts of the 'facets' gem. The idea: because it is known to cause conflicts with the 'activesupport' gem (seen as late as today), let us deprecate all of this as far as our platform is concerned, so we can get see where it is being used and eventually get rid of the whole thing, in a controlled manner.

I know that global variables are bad. The idea here is to set a flag to avoid more warnings than is needed. I.e. if the user/an app is loading 'facets', we only need to produce a single warning for that - no need to warn multiple times. Maybe that's something we could live with, but... I'd rather not spend more time on this now since it's legacy cruft (from the perspective of our platform) so I don't think 100% pureness is needed in this case.